feat(App): integrate React Query for data fetching, refactor project loading logic, and update screen imports for better organization
