+++ app/screens/index.ts
@@ -1,15 +1,14 @@
 export * from "./WelcomeScreen"
 export * from "./Project/CreateProject"
 export * from "./Project/CreateProjectForm"
-export * from "./LoginScreen"
+export * from "./Auth/LoginScreen"
 export * from "./DemoCommunityScreen"
 export * from "./DemoDebugScreen"
 export * from "./DemoShowroomScreen/DemoShowroomScreen"
 export * from "./Project/ProjectsScreen"
-export * from "./ChatScreen"
+export * from "./Chat/ChatScreen"
 export * from "./Project/ProjectSuccess"
 export * from "./DemoPodcastListScreen"
-
 export * from "./LoadingScreen"
-
-export * from "./ErrorScreen/ErrorBoundary"
+export * from "./Error/ErrorBoundary"
+export * from "./Error/ErrorDetails"
